fix(onboarding): verify local gateway readiness before adopting an assistant
The adopt path skipped all readiness checks on the assumption that the
hatching screen had just confirmed the local gateway's /readyz. A
session-based adopt (refresh or direct visit with no query string) never
passed through that screen, and a cached gateway token proves nothing about
the gateway process still being alive — the flow could mark a dead gateway
ready and fail later with worse errors.

Adopt now polls the local gateway's /readyz (gateway + upstream daemon)
before settling ready, sharing one probe helper with the hatching screen.
The assistant-scoped healthz SDK call remains managed-path-only since it
doesn't resolve against a local gateway.

Addresses the Codex P2 review finding on this PR.
